fix dependecies
